Added upstart check in case this is ran twice and added instructions near the end.

+++ support/install_flowbat_local_ubuntu.sh
@@ -104,20 +104,23 @@ exec sudo -u $USER meteor --port 1800 run --settings /home/$USER/opt/FlowBAT/set
 end script
 EOF
 
-read -p "$(tput setaf 3)Do you wish to have FlowBAT start on boot?$(tput sgr0)" -n 1 -r
-        echo
-        if [[ $REPLY =~ ^[Yy]$ ]]; then
-                sudo cp flowbat.conf /etc/init/
-	else
-		echo "$(tput setaf 2)For future reference, move flowbat.conf to /etc/init/ if you would like to have FlowBAT start on boot.$(tput sgr0)".
-	fi
+if [ ! -f /etc/init/flowbat.conf ]; then
+	read -p "$(tput setaf 3)Do you wish to have FlowBAT start on boot?$(tput sgr0)" -n 1 -r
+        	echo
+	        if [[ $REPLY =~ ^[Yy]$ ]]; then
+	                sudo cp flowbat.conf /etc/init/
+		else
+			echo "$(tput setaf 2)For future reference, move flowbat.conf to /etc/init/ if you would like to have FlowBAT start on boot.$(tput sgr0)".
+		fi
+fi
 
 sudo chown -R "$USER":"$USER" /home/"$USER"/
+
+echo -e "$(tput setaf 2)To manually run FlowBAT, cd to /home/$USER/opt/FlowBAT and run:"
+echo -e 'meteor --port 1800 run --settings settings/dev.json "$@"'
+echo -e "$(tput sgr0)"
 
 echo "$(tput setaf 2)Attempting startup. This may take a few minutes if it is the first time. Press ctrl+c to stop FlowBAT after the application says it is running or proceed to 127.0.0.1:1800 in a browser.$(tput sgr0)"
 
 meteor --port 1800 run --settings settings/dev.json "$@"
 
-echo -e "$(tput setaf 2)To manually run FlowBAT again, cd to /home/$USER/opt/FlowBAT and run:"
-echo -e 'meteor --port 1800 run --settings settings/dev.json "$@"'
-echo -e "$(tput sgr0)"
